Move components to Components folder. Strings to strings folder.
