+++ frontend/Components/Card.jsx
@@ -1,9 +1,9 @@
 // @flow
 import * as React from 'react';
-import type { Availability } from './Availability';
-import type { Locale } from './Locale';
-import type { Location } from './Location';
-import strings from './Card.json';
+import type { Availability } from '../Types/Availability';
+import type { Locale } from '../Types/Locale';
+import type { Location } from '../Types/Location';
+import strings from '../Strings/Card.json';
 
 function getBadgeClassname(availability: Availability): string {
   switch (availability) {
