+++ frontend/Components/G0vbar.jsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 // $FlowIgnore[cannot-resolve-module] This is how ParcelJS imports assets
 import favicon from 'url:../assets/g0v.png'; // eslint-disable-line
 
-import type { Locale } from './Locale';
+import type { Locale } from '../Types/Locale';
 
 const g0vbar = {
   background: '#343A40',
