+++ frontend/Strings/Card.json
