+++ frontend/Types/Availability.js
